+++ src/eval.rs
@@ -451,6 +451,8 @@ impl<'a> Evaluator<'a> {
                     (Value::List(_), "sort") => Some(&stdlib::LIST_SORT),
                     (Value::List(_), "sort_by") => Some(&stdlib::LIST_SORT_BY),
                     (Value::List(_), "sum") => Some(&stdlib::LIST_SUM),
+                    (Value::List(_), "to_set_dedup") => Some(&stdlib::LIST_TO_SET_DEDUP),
+                    (Value::List(_), "to_set_unique") => Some(&stdlib::LIST_TO_SET_UNIQUE),
 
                     (Value::Set(_), "all") => Some(&stdlib::SET_ALL),
                     (Value::Set(_), "any") => Some(&stdlib::SET_ANY),